+++ Samples/iOS-Swift/iOS-Swift/Profiling/BenchmarkingViewController.swift
@@ -1,0 +1,129 @@
+import Sentry
+import UIKit
+
+class BenchmarkingViewController: UIViewController, UITextFieldDelegate {
+
+    @IBOutlet weak var workThreadLabel: UILabel!
+    @IBOutlet weak var workIntensityFactorLabel: UILabel!
+    
+    @IBOutlet weak var workThreadSlider: UISlider!
+    @IBOutlet weak var workIntervalSlider: UISlider!
+    
+    @IBOutlet weak var maxThreadsTextField: UITextField!
+    @IBOutlet weak var minThreadsTextField: UITextField!
+    @IBOutlet weak var minWorkIntensityTextField: UITextField!
+    @IBOutlet weak var maxWorkIntensityTextField: UITextField!
+    @IBOutlet weak var valueTextField: UITextField!
+        
+    @IBOutlet weak var dsnView: UIView!
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        minWorkIntensityTextField.text = String(defaultLongestIntervalMicros)
+        maxWorkIntensityTextField.text = String(1)
+        minThreadsTextField.text = String(4)
+        maxThreadsTextField.text = String(50)
+        [maxThreadsTextField, minThreadsTextField, minWorkIntensityTextField, maxWorkIntensityTextField].forEach {
+            $0?.delegate = self
+        }
+    }
+    
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        SentrySDK.reportFullyDisplayed()
+        
+        addDSNDisplay(self, vcview: dsnView)
+    }
+    
+    @IBAction func startBenchmark(_ sender: UIButton) {
+        highlightButton(sender)
+        SentryBenchmarking.startBenchmark()
+    }
+    
+    @IBAction func stopBenchmark(_ sender: UIButton) {
+        highlightButton(sender)
+        guard let value = SentryBenchmarking.stopBenchmark() else {
+            let alert = UIAlertController(title: "Benchmark Error", message: "No benchmark result available.", preferredStyle: .alert)
+            alert.addAction(UIAlertAction(title: "OK", style: .default))
+            self.present(alert, animated: false)
+            print("[iOS-Swift] [debug] [ProfilingViewController] no benchmark result returned")
+            return
+        }
+        valueTextField.isHidden = false
+        valueTextField.text = value
+        print("[iOS-Swift] [debug] [ProfilingViewController] benchmarking results:\n\(value)")
+    }
+    
+    @IBAction func startCPUWork(_ sender: UIButton) {
+        highlightButton(sender)
+        _adjustWorkThreadsToCurrentRequirement()
+    }
+    
+    @IBAction func minWorkThreadCountChanged(_ sender: Any) {
+        _adjustWorkThreadsToCurrentRequirement()
+    }
+    
+    @IBAction func workThreadSliderChanged(_ sender: Any) {
+        _adjustWorkThreadsToCurrentRequirement()
+    }
+    
+    @IBAction func maxWorkThreadCountChanged(_ sender: Any) {
+        _adjustWorkThreadsToCurrentRequirement()
+    }
+    
+    @IBAction func endCPUWork(_ sender: UIButton) {
+        highlightButton(sender)
+        cpuWorkthreads.forEach { $0.cancel() }
+        cpuWorkthreads.removeAll()
+    }
+    
+    @IBAction func minWorkIntervalChanged(_ sender: Any) {
+        _adjustWorkIntervalToCurrentRequirements()
+    }
+    
+    @IBAction func workIntensityChanged(_ sender: UISlider) {
+        _adjustWorkIntervalToCurrentRequirements()
+    }
+    
+    @IBAction func maxWorkIntervalChanged(_ sender: Any) {
+        _adjustWorkIntervalToCurrentRequirements()
+    }
+    
+    @IBAction func bgBrightnessChanged(_ sender: UISlider) {
+        view.backgroundColor = .init(white: CGFloat(sender.value), alpha: 1)
+    }
+        
+    // MARK: UITextFieldDelegate
+    
+    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+        textField.resignFirstResponder()
+    }
+    
+    func _adjustWorkThreadsToCurrentRequirement() {
+        let maxThreads = (maxThreadsTextField.text! as NSString).integerValue
+        let minThreads = (minThreadsTextField.text! as NSString).integerValue
+        let requiredThreads = _projectedRange(factor: workThreadSlider.value, min: minThreads, max: maxThreads)
+        let diff = requiredThreads - cpuWorkthreads.count
+        if diff == 0 {
+            return
+        } else if diff > 0 {
+            for _ in 0 ..< diff {
+                let thread = WorkThread()
+                thread.qualityOfService = .userInteractive
+                cpuWorkthreads.insert(thread)
+                thread.start()
+            }
+        } else {
+            let absDiff = abs(diff)
+            for _ in 0 ..< absDiff {
+                let thread = cpuWorkthreads.removeFirst()
+                thread.cancel()
+            }
+        }
+    }
+
+    private func _adjustWorkIntervalToCurrentRequirements() {
+        let minInterval = (minWorkIntensityTextField.text! as NSString).integerValue
+        let maxInterval = (maxWorkIntensityTextField.text! as NSString).integerValue
+        workIntervalMicros = UInt32(_projectedRange(factor: workIntervalSlider.value, min: minInterval, max: maxInterval))
+    }
+}
