+++ Samples/iOS-Swift/iOS-Swift/Profiling/ProfilingViewController.swift
@@ -2,33 +2,17 @@ import Sentry
 import UIKit
 
 class ProfilingViewController: UIViewController, UITextFieldDelegate {
-    
-    @IBOutlet weak var workThreadLabel: UILabel!
-    @IBOutlet weak var workIntensityFactorLabel: UILabel!
-    
-    @IBOutlet weak var workThreadSlider: UISlider!
-    @IBOutlet weak var workIntervalSlider: UISlider!
-    
-    @IBOutlet weak var maxThreadsTextField: UITextField!
-    @IBOutlet weak var minThreadsTextField: UITextField!
-    @IBOutlet weak var minWorkIntensityTextField: UITextField!
-    @IBOutlet weak var maxWorkIntensityTextField: UITextField!
-    @IBOutlet weak var valueTextField: UITextField!
-    
     @IBOutlet weak var launchProfilingMarkerFileCheckButton: UIButton!
     @IBOutlet weak var profilingUITestDataMarshalingTextField: UITextField!
     @IBOutlet weak var profilingUITestDataMarshalingStatus: UILabel!
         
+    @IBOutlet weak var sampleRateField: UITextField!
+    @IBOutlet weak var samplerSwitch: UISwitch!
+    @IBOutlet weak var samplerValueField: UILabel!
     @IBOutlet weak var dsnView: UIView!
+
     override func viewDidLoad() {
         super.viewDidLoad()
-        minWorkIntensityTextField.text = String(defaultLongestIntervalMicros)
-        maxWorkIntensityTextField.text = String(1)
-        minThreadsTextField.text = String(4)
-        maxThreadsTextField.text = String(50)
-        [maxThreadsTextField, minThreadsTextField, minWorkIntensityTextField, maxWorkIntensityTextField].forEach {
-            $0?.delegate = self
-        }
         profilingUITestDataMarshalingTextField.accessibilityLabel = "io.sentry.ui-tests.profile-marshaling-text-field"
         launchProfilingMarkerFileCheckButton.accessibilityLabel = "io.sentry.ui-tests.app-launch-profile-marker-file-button"
     }
@@ -38,66 +22,25 @@ class ProfilingViewController: UIViewController, UITextFieldDelegate {
         SentrySDK.reportFullyDisplayed()
         
         addDSNDisplay(self, vcview: dsnView)
-    }
-    
-    @IBAction func startBenchmark(_ sender: UIButton) {
-        highlightButton(sender)
-        SentryBenchmarking.startBenchmark()
-    }
-    
-    @IBAction func stopBenchmark(_ sender: UIButton) {
-        highlightButton(sender)
-        guard let value = SentryBenchmarking.stopBenchmark() else {
-            let alert = UIAlertController(title: "Benchmark Error", message: "No benchmark result available.", preferredStyle: .alert)
-            alert.addAction(UIAlertAction(title: "OK", style: .default))
-            self.present(alert, animated: false)
-            print("[iOS-Swift] [debug] [ProfilingViewController] no benchmark result returned")
-            return
+
+        let options = SentrySDK.currentHub().getClient()?.options
+
+        if let sampleRate = options?.profilesSampleRate {
+            sampleRateField.text = String(format: "%.2f", sampleRate.floatValue)
+        } else {
+            sampleRateField.text = "nil"
         }
-        valueTextField.isHidden = false
-        valueTextField.text = value
-        print("[iOS-Swift] [debug] [ProfilingViewController] benchmarking results:\n\(value)")
+
+        samplerSwitch.isOn = options?.profilesSampler != nil
+        if options?.profilesSampler != nil {
+            if let samplerValue = SentrySDKTestConfiguration.Profiling.getSamplerValue() {
+                samplerValueField.text = String(format: "%.2f", samplerValue)
+            } else {
+                samplerValueField.text = "dynamic"
+            }
+        }
     }
-    
-    @IBAction func startCPUWork(_ sender: UIButton) {
-        highlightButton(sender)
-        _adjustWorkThreadsToCurrentRequirement()
-    }
-    
-    @IBAction func minWorkThreadCountChanged(_ sender: Any) {
-        _adjustWorkThreadsToCurrentRequirement()
-    }
-    
-    @IBAction func workThreadSliderChanged(_ sender: Any) {
-        _adjustWorkThreadsToCurrentRequirement()
-    }
-    
-    @IBAction func maxWorkThreadCountChanged(_ sender: Any) {
-        _adjustWorkThreadsToCurrentRequirement()
-    }
-    
-    @IBAction func endCPUWork(_ sender: UIButton) {
-        highlightButton(sender)
-        cpuWorkthreads.forEach { $0.cancel() }
-        cpuWorkthreads.removeAll()
-    }
-    
-    @IBAction func minWorkIntervalChanged(_ sender: Any) {
-        _adjustWorkIntervalToCurrentRequirements()
-    }
-    
-    @IBAction func workIntensityChanged(_ sender: UISlider) {
-        _adjustWorkIntervalToCurrentRequirements()
-    }
-    
-    @IBAction func maxWorkIntervalChanged(_ sender: Any) {
-        _adjustWorkIntervalToCurrentRequirements()
-    }
-    
-    @IBAction func bgBrightnessChanged(_ sender: UISlider) {
-        view.backgroundColor = .init(white: CGFloat(sender.value), alpha: 1)
-    }
-    
+
     @IBAction func checkLaunchProfilingMarkerFile(_ sender: Any) {
         let launchProfileMarkerPath = ((NSSearchPathForDirectoriesInDomains(.cachesDirectory, .userDomainMask, true).first! as NSString).appendingPathComponent("io.sentry") as NSString).appendingPathComponent("profileLaunch")
         if FileManager.default.fileExists(atPath: launchProfileMarkerPath) {
@@ -128,7 +71,36 @@ class ProfilingViewController: UIViewController, UITextFieldDelegate {
             handleContents(file: file)
         }
     }
-        
+
+    @IBAction func sampleRateEdited(_ sender: UITextField) {
+        guard let text = sender.text, !text.isEmpty else {
+            UserDefaults.standard.removeObject(forKey: SentrySDKTestConfiguration.Profiling.Key.sampleRate.rawValue)
+            return
+        }
+
+        let rate = (text as NSString).floatValue
+        UserDefaults.standard.set(rate, forKey: SentrySDKTestConfiguration.Profiling.Key.sampleRate.rawValue)
+    }
+
+    @IBAction func samplerSwitchToggled(_ sender: UISwitch) {
+        if sender.isOn {
+            if let text = samplerValueField.text {
+                let rate = (text as NSString).floatValue
+                UserDefaults.standard.set(rate, forKey: SentrySDKTestConfiguration.Profiling.Key.samplerValue.rawValue)
+                samplerValueField.text = String(format: "%.2f", rate)
+            }
+        } else {
+            UserDefaults.standard.removeObject(forKey: SentrySDKTestConfiguration.Profiling.Key.samplerValue.rawValue)
+        }
+    }
+    
+    @IBAction func samplerValueEdited(_ sender: UITextField) {
+        if let text = sender.text {
+            let rate = (text as NSString).floatValue
+            UserDefaults.standard.set(rate, forKey: SentrySDKTestConfiguration.Profiling.Key.samplerValue.rawValue)
+        }
+    }
+
     // MARK: UITextFieldDelegate
     
     func textFieldShouldReturn(_ textField: UITextField) -> Bool {
@@ -181,34 +153,5 @@ class ProfilingViewController: UIViewController, UITextFieldDelegate {
             profilingUITestDataMarshalingTextField.text = "<empty>"
             profilingUITestDataMarshalingStatus.text = "❌"
         }
-    }
-    
-    func _adjustWorkThreadsToCurrentRequirement() {
-        let maxThreads = (maxThreadsTextField.text! as NSString).integerValue
-        let minThreads = (minThreadsTextField.text! as NSString).integerValue
-        let requiredThreads = _projectedRange(factor: workThreadSlider.value, min: minThreads, max: maxThreads)
-        let diff = requiredThreads - cpuWorkthreads.count
-        if diff == 0 {
-            return
-        } else if diff > 0 {
-            for _ in 0 ..< diff {
-                let thread = WorkThread()
-                thread.qualityOfService = .userInteractive
-                cpuWorkthreads.insert(thread)
-                thread.start()
-            }
-        } else {
-            let absDiff = abs(diff)
-            for _ in 0 ..< absDiff {
-                let thread = cpuWorkthreads.removeFirst()
-                thread.cancel()
-            }
-        }
-    }
-
-    private func _adjustWorkIntervalToCurrentRequirements() {
-        let minInterval = (minWorkIntensityTextField.text! as NSString).integerValue
-        let maxInterval = (maxWorkIntensityTextField.text! as NSString).integerValue
-        workIntervalMicros = UInt32(_projectedRange(factor: workIntervalSlider.value, min: minInterval, max: maxInterval))
     }
 }
